fix: refactor telegraf version
