+++ plugins/inputs/internal/internal_test.go
@@ -21,13 +21,14 @@ func TestSelfPlugin(t *testing.T) {
 	stat.Incr(1)
 	stat.Incr(2)
 	require.NoError(t, s.Gather(acc))
+
 	acc.AssertContainsTaggedFields(t, "internal_mytest",
 		map[string]interface{}{
 			"test": int64(3),
 		},
 		map[string]string{
 			"test":    "foo",
-			"version": "",
+			"version": "unknown",
 		},
 	)
 	acc.ClearMetrics()
@@ -41,7 +42,7 @@ func TestSelfPlugin(t *testing.T) {
 		},
 		map[string]string{
 			"test":    "foo",
-			"version": "",
+			"version": "unknown",
 		},
 	)
 	acc.ClearMetrics()
@@ -59,7 +60,7 @@ func TestSelfPlugin(t *testing.T) {
 		},
 		map[string]string{
 			"test":    "foo",
-			"version": "",
+			"version": "unknown",
 		},
 	)
 }
